Cryptsetup processes are now only killed when the device is unlocked via SSH, since that is the only situation that multiple cryptsetup processes are running.
The code to setup the initrd network has been made configurable.
This removes the need to configure the network via an invalid kernel parameter,
and it makes it flexible enough to handle the many use cases people may have,
such as using WLAN.

An additional option has been added to add the necessary setup dependencies,
such as the wpa_applicant binary in the case of WLAN.

Mounting /dev/pts in order to allow SSH connections caused problems with
some terminals. Which has been resolved by lazily unmounting it.